ref(server): Create an entry point for event metric extraction
